[wip] extract hash code to module
